video.Contents: Handle portrait orientation.
Currently, video.Contents is not handling screenshots taken in dru
correctly. This is because we rotate the screenshot when the width is
less than the height in order to handle rotated screens like in krane.
The problem is that in dru (at least in the lab), the display
orientation, portrait, is the same as the screenshot orientation, so the
screenshot should not be rotated. In krane, the display orientation,
landscape, is not the same as the screenshot orientation, portrait, so
the screenshot should be rotated.

To address this, prior to taking the screenshot, we rotate the display
to get to landscape orientation if not already there. That way we know
that rotating portrait screenshots is correct.

This doesn't sound perfect, but it works on the devices I tested (eve,
dru, and krane). For example, it's not obvious why we need to rotate 270
degrees instead of 90 degrees. I found this by trial and error. I'll
need to see the results across all the devices in the lab.

BUG=b:159158467
TEST=h264_360p_exotic_crop_composited_hw on eve, dru, and krane.

Change-Id: I8ce6a23a59a937bb6cc591541beb27822b48d3ea