tast-test: Stabilize HotseatScrollPerf.clamshell_mode
The HotseatScrollPref test performs shelf scroll by clicking the shelf
scroll arrow buttons, whose bounds are fetched from the autotestPrivate
API before scrolling. The first step is to ensure tablet mode, which may
update the shelf bounds, and thus the shelf arrow bounds. The test did
not wait for the UI to stabilize after tablet mode change, which made it
possible for the test to try scrolling by clicking obsolete arrow
bounds, and failing.

This change adds a wait for the location changes to stabilize after
tablet mode change, which should ensure the shelf arrow bounds are
stable when the scroll test starts.

BUG=chromium:1142206
TEST=Ran the test on kukui

Change-Id: I79b94c5b5bf3a679b57a9e73f59dfd148d37afe9