tast-test: Stabilize HotseatScrollPerf.clamshell_mode
The HotseatScrollPref test performs shelf scroll by clicking the shelf
scroll arrow buttons, whose bounds are fetched from the autotestPrivate
API before scrolling.
After CL:2436648 tests clear all notifications as part of resetting the
test state - this impacts the shelf bounds if notifications are shown,
which means that the shelf bounds might be changing as the shelf scroll
test is starting.

This change adds a wait for the location changes to stabilize before the
scroll test starts, ensuring that the shelf arrow bounds used by the
scroll test are not stale.

BUG=chromium:1142206
TEST=Ran the test on kukui

Change-Id: I79b94c5b5bf3a679b57a9e73f59dfd148d37afe9