wificell: Add Start, Stop interface for frame sender.
In some tests, we want to spawn a long running background frame sender,
Start/Stop interface will be more convenient for the usage instead of
the synchronous Send interface. Also, it can abort the command with
pkill in case the process is not properly killed by the default
Abort when context closed due to crbug.com/1030635.

BUG=b:171763917
TEST=tast run $DUT wifi.CSALeaveChannel wifi.FunctionalAfterCSA*

Change-Id: Ice2c07aba9d31c84f48c18f6a86286dfb24292b4