selinux: Remove rootfs parts from SELinuxFilesSystem tast tests
Directories that are part of the rootfs shouldn't be tested
against SELinux policy change in here. /bin /sbin /usr/bin /usr/sbin
and /etc are removed from this tast test. SELinux policy context for
these directories should be tested in image test and will be added
there.

This CL aims to reduce the unnecessary testing steps in the SELinux
tast tests, as part of the effort of GreenStainlessTast. We want
SeLinux tast tests focus on smaller subset of more meaningful tests
that fail only when they really fail.

BUG=chromium:1129062
TEST=run on DUT

Change-Id: Iefcf7f95c7fa047a2e9a0d66af295a3d36e61d38