selinux: Remove rootfs parts from SELinuxFilesSystem tast tests
Directories that are part of the rootfs don't need to be tested
against SELinux policy change. Therefore /bin /sbin /usr/bin /usr/sbin
and /etc are removed from this tast test.

Within the effort of GreenStainlessTast we want SeLinux tast tests focus
on smaller subset of more meaningful tests that fail only when they
really fail. This CL is aims to reduce the unnecessary testing steps.

BUG=chromium:1129062
TEST=run on DUT

Change-Id: Iefcf7f95c7fa047a2e9a0d66af295a3d36e61d38