Remove settings page from ui.IdlePerf'*'
I'm not sure why settings page was added to this test, but for
the essence of this test case, it does not make much sense to
have that. Simply keeping idle should be just good enough for
this test case.

Also this test removes 'Timeout'; it won't take that long time.

BUG=none
TEST=tast run DUT ui.IdlePerf'*'

Change-Id: I4632c2f0361e88c12e787e2ac947175c09e3ac7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2504360
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>