Uprev cups-filters to 1.28.5
Change the ghostscript tests output file name to the golden file name to
be consistent with the other tests and make it easier to update the
golden files.

Modify lpprint.CleanPSContents() to remove all postscript comments (this
is the easy way to handle the multiline gs Invocation comment).

Update the golden files for tests affected by the uprev.

BUG=none
TEST=tast run localhost:9222 printer.*

Cq-Depend: chromium:2500412
Change-Id: I66151b4b51e21a582447e7ddc936fa65e7328b05