wifi: OptionalDHCPProperties: Capture L3 packets on interface of AP.
As the test only cares DHCP requests but not the 802.11 frames, we can
capture the packets directly on the interface of AP on router. With this
change, it'll be unlikely to drop any target packets as DHCP server on
the same interface works, and make the test more stable.

BUG=b:162395242
TEST=tast run $DUT wifi.OptionalDHCPProperties

Change-Id: I0e42a4c92e81da65547997dd97d7b703c0c8c2bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2505038
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>