tast-tests: update platform.CrosHealthdProbeCPUInfo
The tast test needs to be made aware of the new temperature channel
data coming from cros_healthd. This data doesn't need to be
verified, since it is optional. However, the logic for identifying
physical CPUs needs to be modified to check for the new line of
headers. If they're found, the current physical CPU must be complete
and should be verified.

BUG=chromium:1131602
TEST=tast run --build $DUT_IP platform.CrosHealthdProbeCPUInfo

Cq-Depend: chromium:2490581
Change-Id: Ia43da30a8a94d765086a4a746c3a8664aeb644f9