Improve resizing disk during installation
Allow the installer to use a disk size smaller than the specified value.
Some of the devices in the lab have very small disk size. Therefore,
forcing to resize the disk to be larger than a specified value could
cause flakiness. This cl allows the installer to use the maximum disk
size if the specified value is larger than it.

BUG=chromium:1143939
TEST=On Hatch

Change-Id: I35e2143c62b6861a604b4d2c4a9d0564ac117792
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2507351
Commit-Queue: Judy Wu <jinrongwu@google.com>
Tested-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>