Fix failure of HomeDirectoryCreateFile
The error is Failed to test creating files in Linux files: creation of
folder "test_folder" did not propogate to container: failed to find
test_folder in container. It might because that the new folder was not
created with context menu. Changing to press Ctrl+E to create the new
folder.

BUG=chromium:1144007
TEST=On Hatch

Change-Id: I36aae60756c4de31ad3b70ce3d74618b96a8b729