wificell: Add option for TestFixture to not stop UI.
In normal WiFi tests, we always stop UI as it can cause some
interference (e.g. request scans). Now in some tests, it will have to
interact with UI for some settings and also verify basic WiFi
functionality (connect, ping ...). In these tests, the interference
from UI usually does no harm. Add an option on TestFixture for this,
and we can even share the setup for these tests with another
Precondition in the future.

BUG=b:161902990
TEST=`tast run $DUT wifi.SimpleConnect.80211g` and checks UI is still
stopped in this test.

Change-Id: If50833e4b01b952972d734a81ba99e4424a94104