video.Contents: Increase stencil size.
This CL increases the size of the stencil used to test the color of the
inner corners of the video screenshot. This is to increase the tolerance
for blending artifacts. Notably, it seems that in bob, displaying the
1080p video was causing the video to get downscaled and we were sampling
pixels in the middle of the blending artifacts, so the color distance
was large.

BUG=b:159158467
TEST=video.Contents.h264_1080p_composited_hw on bob

Change-Id: I72adacbe14c09df127b431c84af48cc5f17229fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2515106
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>