Fix 10% flake chance for tast.printer.AddHpPrinter
Previously the test would flake if the job-start-time milliseconds were
less than a hundred (2 characters instead of 3). This CL fixes that by
removing the PDF xref table before computing the diff.

Also the ghostscript test backend used for testing ghostscript-related
print filters is changed to use the lpprint CleanPSContents() function
in case any of the filters include the job-start-time in the PDF output.

BUG=none
TEST=manual

Change-Id: I12a34202a298f2537637892eb67bdecc88ec6c57