tast-test: Stabilize shelf bounds before HotseatAnimation
Attempt to make HotseatAnimation tests more stable - CL:2473557 removed
a wait for CPU cooldown at the start of the test, changing the timing of
when hotseat animation starts, making it more likely for it to start
while shelf transition for entering tablet mode was still in progress.

This adds a wait for shelf bounds to stabilize after tablet mode
transition before the hotseat animation starts.

BUG=chromium:1143980
TEST=HotseatAnimation tests pass on kukui

Change-Id: I65793adbc1fe1d8d52a9d3b461873567582bfb1f