tast-tests: Record who is sending SIGKILL to the bundle
This is the revised version of https://crrev.com/c/2513678 which uses
pre/post-runhooks.

As the issue is not reproducible locally and it appears the sender is
not the OOM killer, it is useful to record the kill syscalls to the
bundle process in audit.log for further investigation.

BUG=b:169712138
TEST=sending sigkill to the bundle recorded in audit.log

Cq-Depend: chromium:2513683
Change-Id: Ie988a0c4d1def10e64bf2fd1e7d38a75a21ba5dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2517264
Tested-by: Tetsui Ohkubo <tetsui@chromium.org>
Commit-Queue: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>