tast-tests: crostini.VmcExtraDisk: Use shorter VM name
The VM name passed to vmc will be encoded in the name of unix domain
sockets. With the current long VM name, the length of socket path exceeds
107, which is defined as the length of `sun_path` in `struct sockaddr_un`.
This will make a VM fail to startup.
So, we must use a shorter VM name.n

BUG=chromium:1144639
TEST=`tast run crostini.VmcExtraDisk.*` passed on kukui

Change-Id: Icc9f09abdd53a4be999d5cc246c073b49f43d244
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2517273
Tested-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>