firmware.GBBFlags: Log error from initial Get
Due to a typo in the firmware.GBBFlags test, if the initial GetgBBFlags
fails, the error is not logged.  This causes some failures on Tastboard
to be difficult to debug: see tastboard/test/firmware.GBBFlags.

This change fixes that typo.

BUG=b:172463577
TEST=firmware.GBBFlags
TEST=firmware.GBBFlags with artificially induced error

Change-Id: Ie42381b2a387ed02a025868db3afd2d407406c89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2519485
Tested-by: Greg Edelston <gredelston@google.com>
Auto-Submit: Greg Edelston <gredelston@google.com>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>