tast-tests: Test that fast ink is working as evidenced by an overlay
For now, it will test only in clamshell mode, and only with the window
states Normal, Maximized, and Fullscreen. The test fails if it tries a
snapped window state in clamshell mode, or any window state in tablet
mode. The test will be expanded as these cases are addressed.

BUG=chromium:1070307
TEST=graphics.FastInk

Change-Id: Iea6bcb700f058c3bd5f5fdf2e289ed35ecb4b527
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2565383
Tested-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>