tast-tests/video: move ChromePIPEnergyAndPower to ui/
video.ChromePIPEnergyAndPower are failing and their monitoring/triaging
is suboptimal due to the folder where they are. This CL moves this
file to the ui/ folder.

The pip.html file used by this test is renamed to pip_video.html to
avoid colliding with the existing ui/data/pip.html.

BUG=b:171750289
TEST=tast run DUT ui.ChromePIPEnergyAndPower.* on celes

Change-Id: I083426da9462cd6e03e986c7a278a1bab7afb406