Stablize ui.LauncherSearchBuiltInApps.
It is failing on tablet devices a lot just to open the launcher.
Screenshots show that the launcher is not opened. It is currently using
a keyboard emulator to open the launcher. Previously it went through the
UI but tablet mode doesn't have the shelf launcher button.

Seewai added a function for the launcher recently that does a lot more
checking for things in tablet mode so let's just reuse that.

BUG=chromium:1145809
TEST=ui.LauncherSearchBuiltInApps

Change-Id: I00a6a832ff662478c4152513d15fccb2dfd95887