Stabilize ui.WindowCycle
This test recently started failing 100% on many boards. The main issue
is that the test was not properly waiting for Files and Settings to
load. If window cycling starts before one of their UIs has rendered,
that app will pop up on top once it fully loads. The test tried to
account for this using what's available in ash/wm.go, but it's much more
reliable to wait for something specific in the UI to load. The test will
now wait for some specific UI elements to load in Files and Settings
before moving on to make sure the window order is stable before we start
cycling windows.

BUG=chromium:1081688
TEST=tast run <ip> ui.WindowCycle
TEST=./repeat-test.sh -s -r 500 chromeos6-row14-rack22-host1.cros
ui.WindowCycle (test passed 500 consecutive runs on an octopus/bluebird
device, which has a 100% failure rate in the lab on recent builds)

Change-Id: I220bb648eb82b649cee4e4d94de079f51d61b69c