tast-tests: fix tconn crash in VirtualKeyboardDeadKeys
Description: tconn crashes in VirtualKeyboardDeadKeys test. The reason is because it uses the same precondition as VirtualKeyboardAccessibility.
The crash happens when reset test environment. So the fix is in precondition and  VirtualKeyboardAccessibility.

Root cause: In VirtualKeyboardAccessibility test, it presses functional keys on a11y virtual keyboard. These are sticky keys, status persists and it causes tconn crash when hide virtual keyboard and reload.

Solution:
1. Change press keys to check keys present.
2. Enhance precondition to hide virtual keyboard if it is on screen.

BUG=chromium:1145841
TEST=inputs.*

Change-Id: I2a7617031c4768bd15820e584af694a125632cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2521979
Tested-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Jun Sheng <shengjun@chromium.org>