tast-tests: stabilize InputMethodManagement test
From the error log in https://tastboard.corp.google.com/test/ui.InputMethodManagement, the node is shown in screenshot but not a11y tree. And it only happens a couple of times out of 100+ runs. So it might just because of low-end board performance issue.

BUG=b:172024570
TEST=ui.InputMethodManagement

Change-Id: I395f13225dd3966a38392fe559fe76b390820451
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2522712
Tested-by: Jun Sheng <shengjun@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Jun Sheng <shengjun@chromium.org>