tast-tests: Test Drag Drop from Files app
Verify that a file dragged from the Files app can be dropped onto a
target on a separate Chrome app. The Chrome app was modified from the
arc.DragDrop test to listen for a drop event, interrogate the dropped
file and pass it back via the document.window.title.

BUG=chromium:1145322
TEST=tast run <DUT-IP> filemanager.DragDrop

Change-Id: Id32c92596bb47929841b8cc078df7df29f324b27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2522542
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@google.com>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>