media: Disable message center notifications.
This CL adds an extra argument for the Chrome media preconditions so
that message center notifications are suppressed. This is particularly
important for the video.Contents tests because they take a screenshot
and check a few pixels. If the screenshot contains notifications, the
test will be flaky (we've actually seen this situation). In general,
notifications are not expected to be useful for our tests.

BUG=b:159158467
TEST=--suppress-message-center-popups hides notifications in
     cryptohome.ShowLowDiskSpaceNotification

Change-Id: Iaf9b5322f5398bf02ca397eea417555d403e5047