wifi: Port network_WiFi_ChannelHop to Tast.
Port the test to Tast as wifi.ChannelHop.

In Tast version, the behavior is changed a little because power save is
not always turned off in Tast. When power save is on, it can take the
DUT long time to notice the new AP and reconnect (sometimes ~1 minute).
Triggering active scans can help in this case and it might be more
similar to real use cases (e.g. UI might regularly trigger scans)

BUG=b:172785860
TEST=tast run $DUT wifi.ChannelHop

Change-Id: I5309265ff60879b0f1a73afbbd47617532e65a0d