Fix and reenable unicorn login test.
Test was failing because the buttons no longer have role=button
attribute. Changed to read the data-email attribute instead.
The test used to live in the private repo but it doesn't need to be
there so now the test will be in the public repo.

BUG=chromium:1097949
TEST=unicorn.Login

Change-Id: Id770fe8a23b02b1e3f3762fd3e2dd617bb2a0ec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2523483
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>