tast-tests: Make LauncherDragPerf work with scrolled shelf
LauncherDragPerf test was getting initial drag position using in-screen
bounds of first two icons within the shelf. This assumed that the shelf
was not scrolled - if the shelf is scrolled, the first two shelf icons
may be off-screen, causing the LauncherDragPerf test to fail (as the
initial drag position would not be within the shelf bounds).

This CL updates the logic for calculating initial drag position to use
first two visible app icons - i.e. first two icons that are right of the
left scroll arrow.

BUG=chromium:1144608
TEST=run tast ui.HotseatScrollPerf ui.LauncherDragPerf with and without
CL:2496739 reverted

Change-Id: Ic3214a4dcad4abbefcf4c26f8878769feb89a69e