Run UIAutomator Test Using Tast For Microsoft Outlook app
BUG=b:172087714
TEST=tast run <eve> arcappcompat.MicrosoftOutlook

Tast Test for MicrosoftOutlook app verifies launch app
from the playstore, app is logged in and reached
main activity page of the app. To login to the app it requires Chrome android app to be downloaded and needed to provide permission for Outlook using chrome android app. Also for Clamshell
Mode it verifies Clamshell Full Screen, Minimize
and Restore, Clamshell Resize, Reopen tests and
for Touchview Mode it verifies Minimize and Restore,
Reopen tests.

Cq-Depend: chrome-internal:3387647
Change-Id: Ib6b6f2df4fd29f0c470645e43081d551a7c4a7b8