tast-tests: Wait for home screen in HotseatAnimation WindowCreation
WindowCreation subtest creates an app window while on home launcher
screen to test hotseat transition from shown to hidden state. At the end
of the test, it closes the window,but without waiting for the hotseat
transition to shown state finishes (instead, the window was getting
closed as part of defer()). This means that each new iteration
of the test would likely start when the transition to home screen is
still in progress, not actually testing full transition from shown to
hidden state.

This updates the window creation subtest to explicitly handle window
closure, and adds logic to track metrics for transition from hidden to
shown state.

Also, adjusts WaitForStableShelfBounds to explicitly check whether the
shelf bounds are animating.

BUG=chromium:1143980
TEST=ui.HotseatAnimation* on krane, nautilus

Change-Id: I9c8b402042e7c00ce75074a94bb947e6156011c0