Copy syslog before reboot.
Some crash may happen during the reboot procedure.
Log files right before reboot are useful for debugging in such cases.
Copy syslog before reboot, so that it will not be lost.

There has been the logic in ensureTPMIsReset to do that. This change
partially moves that to the shared path run for every reboot by
DUT.Reboot().

BUG=chromium:1097657
TEST=tast run $DUT power.Reboot

Cq-Depend: chromium:2526701
Change-Id: I189c73602702a4efb99bb61e2e35a64646452216