tast-tests: filemanager wait for Downloads to exist
Some filemanager tests seem to be failing in CQ because the Downloads
folder does not exist/is not yet mounted. This makes sure to wait for
the Downloads folder to exist. Likely we hit this issue because of how
fast automated tests run. I have not managed to reproduce the error
locally, so this is just a guess.

BUG=chromium:1146196, b:157687081
TEST=tast run <eve> filemanager.Smoke filemanager.ImageQuickView

Change-Id: Ibf764dafd39307502badb1fa9676b516792a2fb2