Fix flaky ZipMount.mount_multiple
Before typing the new  password, wait for "Invalid password" to display
on the UI.

The test used to be flaky on slow devices, the reason being that we
were waiting for the input field of the password dialog to be editable,
and to contain the previously typed password. But this is already found
immediately after typing the wrong password.
After typing the wrong password, the password dialog reappears and the
input field is populated with the wrong password that was previously
entered. This is what we want to wait for.

BUG=chromium:407010
TEST=tast run DUT_IP filemanager.ZipMount.mount_multiple

Change-Id: I69bd61a5656630e3fa9081cc1b896b351e627feb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2532484
Reviewed-by: François Degros <fdegros@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>
Tested-by: Jeremie Boulic <jboulic@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>