Add RT22 Tast-test for WM ARC++ R
Implement RT22: resizable/clamshell: split screen.
In RT22 two activities are started, the first one is referred to as
underActivity and the second one as overActivity. At first in
clamshell mode a swipe from the top middle of the screen to the
center is simulated, this will take the overActivity to a snapping
state. Then it will be snapped to the left. UnderActivity will be
snapped to the right by a touch on the center of the right side of
the screen. Then window bounds for each activity will be asserted.
Then the display will be rotated to portrait mode. Left snapped
activity (overActivity) will be snapped to the top and the other one
to the bottom and window bounds will be asserted again. At the end
the display will be rotated back to clamshell and window bounds will
be asserted.
For more info please visit go/arc-wm-r.
BUG=b:140203139
TEST=tast run $DUT arc.WMResizableTablet

Change-Id: I5a40a6cebf202b4d46a9c6b57ebd10a78aceb95b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2536752
Tested-by: Armen Keshishian <armenk@google.com>
Commit-Queue: Armen Keshishian <armenk@google.com>
Reviewed-by: Arthur Hsu <arthurhsu@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>