Add RT22 Tast-test for WM ARC++ R
Implement RT22: resizable/clamshell: split screen.
In RT22 two activities are started, the first one is referred to as
underActivity and the second one as overActivity. Then two different
tests run, one in clamshell mode and the other in portrait. In each
test a swipe from the top middle of the screen to the center is
simulated, this will take the overActivity to a snapping state. Then
depending on the display orientation, it will be snapped to left or
bottom. UnderActivity will be snapped to the opposite side by a touch
event. At the end window bounds for each activity will be asserted
based on their position.
For more info please visit go/arc-wm-r.
BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableClamshel

Change-Id: Iaf7adc5e15f9d6388616bb8c9a90dd14f8ec8bf5