printer/reset_cups_state.go: Add stamp file timeout
In theory, if ResetCups() is called before the first login-prompt-visible
signal is emitted, the stamp file may not exist. In this case, it is
ok to continue after waiting a while (10s) instead of waiting forever
for a non-existent instance of cups-clear-state.sh to finish executing.

BUG=none
TEST=tast run localhost:9222 printer.*

Change-Id: I6f95cc203b9b4aa118ad5708e807e959a0347b15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2533666
Commit-Queue: Pranav Batra <batrapranav@chromium.org>
Tested-by: Pranav Batra <batrapranav@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>