Do not wait for debuggin port file on faillog.
It is expected that Chrome with debuggin port open is running
in the case. If missing, we do not expect a new one is created.
So, let it fail immediately in the case.

Note that there's an edge case that the Chrome with debugging port
is now being opened. Though, we didn't not see such a case in
practice.

BUG=chromium:1144905
TEST=Ran graphics.IGT.kms_addfb_basic on zork. No timeout error is found.

Change-Id: I6fd690c8d3b6f9f4f05036104afd0c5716e768d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2537340
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>