tast-tests: Support RvR testbed in wifi tests (WIP)
Refactor APIface to represent generic API interface and move
current implementation to HostapdAP.
Add cisco.AccessPoint as an alternative implementation
providing support of Cisco APs connected via a Cisco wireless controller.
Change Pre and TestFixture logic to support the same test code in both
configurations (with Gale routers and with Cisco controller) detecting
in runtime which setup is available.
Does not support tests to leverage both Gale and Cisco APs in one setup.

This is a skeleton implementation, to be done:
* implement all of APIface for cisco.AccessPoint
* teardown of cisco.AccessPoint
* move HostapdAP implementation to a separate package

See b/171845870 for details on the setup with Cisco APs.
Primary target is to run roaming tests (RoamNatural, RoamDiagnostics)
in such a setup.

BUG=b:161436986
TEST=tast -verbose run -var attenuator=attenhost -var ciscoctrl=user:passwd:ciscohost duthost wifi.RoamNatural

Change-Id: I789f4f551dc74d38138bf6032504cac5ac3502ce