tast-tests: Refactor APIface to represent generic API interface (WIP)
Prepare Wifi tests to support non-Gale access points.
See b/171845870 for details on a setup with Cisco APs.
Primary target is to run roaming tests (RoamNatural, RoamDiagnostics)
in such a setup.

This is a draft proposal of refactoring.
APIface name is reused (to limit amount of changes) but
it's changed to be an actual interface.
The change does not compile, replace of all occurences of *APIface
to APIface is not included to avoid obfuscating the design.

BUG=b:161436986
TEST=TBD

Change-Id: I789f4f551dc74d38138bf6032504cac5ac3502ce