tast-tests: Fix ui.MeetCUJ.base_case failures on grunt
ui.MeetCUJ.base_case failed on grunt that it couldn't enter the meeting.
After investigation, it is due to the long loading[1] of Meet meeting
that permission bubbles didn't show up on time. Thus it failed to skip
the permission requests because of the timing issues and it failed to
join the meeting after that.
This CL increased the timeout of finding permission bubbles from 10s to
20s, based on what I observed on grunt devices in skylab.

[1] https://drive.google.com/file/d/1dRvB0xGUfTJQOJAyR9Vfr5j5bW_pjBT0/
view?usp=sharing

BUG=chromium:1148002
TEST=tast run -var record=true $DUT ui.MeetCUJ.base_case
Change-Id: I0018dd83e5332ade3fd0052e96a251e44c1a69e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2538094
Tested-by: Yichen Zhou <yichenz@chromium.org>
Commit-Queue: Yichen Zhou <yichenz@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>