Fix notificationcentresmoke.
It is likely the mouse cursor on top of the notification causing it
not to dismiss. I have tested locally by putting the cursor there and
the test fails. And the screenshot doesn't include the cursor matching
what we see in lab failure screenshots.

BUG=chromium:1048435
TEST=ui.NotificationCentreSmoke

Change-Id: I8c4012c1f1bbbd4ae8f568b21b281f9698ee2805
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2538077
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>