tast-tests: Mark failing tests as informational
ui.SigninProfileExtension
ui.ExistingUserLogin
debug.Printer

These tests need to be fixed. The were disabled in the chromium build
but that doesn't have an effect on the ChromeOS build, which is
consistently faling on amd64-generic-full, and thereby blocking the
LKGM uprev because there are more than 10 consecutive failures, and
LKGM will only try 10 fallback versions of the ChromeOS sdk before
giving up.

BUG=chromium:1148527
TEST=Manual testing

Change-Id: Ie3336db2e8f6da20ad3bde3b0ddfc6032db82ec3