Added a test for copying from Files app. tast-tests: Add filemanager.CopyImage test
This test creates a image file in the file manager, tests that copying
between folders within the file manager works, and finally tests that a
file object containing an image is pasted when invoked in a browser
window. This test aims to ensure that pasting in a browser window as
well as between files continues to work for images.

BUG=chromium:827333
TEST=tast run <DUT_IP> filemanager.CopyImage

Change-Id: I1838b3aaab08f9738a69fb0f6bf471855c3aca08