security: Use bind mounts in OpenSSLBlocklist
Thus far, the test relied on an environment variable to point the
system at a different blocklist file for testing. This would make the
test pass even if the actual blocklist file is missing. This change
switches the test to use bind mounts over the actual blocklist file so
it exercises actual production behavior.

BUG=chromium:1149895
TEST=tast run $DUT security.OpenSSLBlocklist

Cq-Depend: chromium:2543390
Change-Id: I55e7f05185af06fd1c1c888efa6aa450edf77fda