Fix misspelling in arc.AccessibilitySpeech.vm
We set buggy output as expected output in arc.AccessibilitySpeech.vm,
so it is failing after the bug is fixed by CL:2528515,

I deleted the wrong field in CL:2522715.
I should delete "Not checked", not "not checked".
This CL is to fix this.

BUG=b:154433831
TEST=arc.AccessibilitySpeech.vm

Change-Id: I85018c575ad9d3f46cca562ab8d704c34b284b37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2545588
Auto-Submit: Saho Kobayashi <sahok@chromium.org>
Tested-by: Saho Kobayashi <sahok@chromium.org>
Commit-Queue: Saho Kobayashi <sahok@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Sara Kato <sarakato@chromium.org>