Add RC22 Tast-test for WM ARC++ R
Implement RC22: resizable/clamshell: split screen.
In RC22 a resizable activity is started, by simulating a mouse drag
event from the Maximize caption button to the left, the activity
snaps to the left. Then the second activity is started and
automatically snapped to the right. Then window bounds compared to
the desired snapped window bounds to make sure both activities are
snapped successfully.

BUG=b:140203139
TEST=tast run $DUT arc.WMResizableClamshell

Change-Id: Iadd989733f52f49bd145096db3be997e475e91b7