tast-tests: Add OnlyStableBoards flag to crostini parameter generation
OnlyStableBoards flag during generation of parameters for Crostini tests
will only generate the _stable variants. This is used in the
filemanager.SMB test as the _unstable board variants are producing a lot
of noise that isn't necessarily beneficial to the core purpose of the
test.

BUG=chromium:1084323
TEST=tast run <DUT-IP> filemanager.SMB.*

Change-Id: Ifcaed0b412d9e9dec279c3973412ac4dd0c3e382
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2547269
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Josh Simmons <simmonsjosh@google.com>
Reviewed-by: Ben Reich <benreich@google.com>
Commit-Queue: Ben Reich <benreich@google.com>