Allow test to deal with notification being focused.
The test passes 100% on some boards and fails 100% on others. Failures
are due to the notification getting focused. When this happens the
notification does not autodismiss.

It never happens when run as the only test. But when certain tests run
before this one, the notification gets focused. I cannot tab away
from the notification because the tabs end up just cycling within the
notification :) So instead I am just going to open the launcher. The
notification will then autodismiss and the test continues.

BUG=chromium:1048435
TEST=ui.DesktopControl ui.LauncherSearchBuiltInApps ui.NotificationCentreSmoke

Change-Id: I511c5140f83e80c062594e82d65180883ee42e94