tast-tests: Stabilize filemanager.DragDrop
The Files app UI is asynchronous and it appears that on certain boards
waiting for the Action bar to show 1 file selected can sometimes not be
enough to ensure the file has successfully been attached to the drag
event.

To overcome this update the window.document.title to ensure we only
start the test once the JavaScript has loaded. Once finished wait for
the files app listbox to stabilize to ensure the item has been
successfully selected.

BUG=chromium:1145322
TEST=./repeat-test.sh -s -r 100 -- {volteer,nami,hatch} filemanager.DragDrop

Change-Id: I5e0d6effd98cd1c2f0880ddfedd7390378ce28bd