policy_tast_struct_creater.py: add parse override for a reference object
When serving the policy ManagedBookmarks in tast the BookmarkType value
will write it's children value as null when it is omitted causing the
policy verification to fail.
Instead of setting children to null when omitted it should be omitted
so the flag omitempty needs to be added to the struct when it is
generated. For that a parse override was added.

BUG=chromium:1150782
TEST=run policy_tast_struct_creator.py to create the defs.go file

Change-Id: Ia05700553968b8c1c281239cff949e6ae5c2a13b