tast: Add tests to firmware_smoke group
These tests don't need to run in the CQ, so to save resources we're
moving them to a separate group that will run nightly in the faft_smoke
suite.
Once we are able to monitor results of these tests, we'll remove them
from the mainline/informational groups.

BUG=b:173440149
TEST=test_that --args="$(SERVO_ARGS)" $HOST tast.firmware-smoke

Change-Id: I616e734c981138d677615a147e9339411ac7e565
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2552666
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>