tast-tests: Port network_WiFi_BT_AntennaCoex to tast
Adds a test that verifies that WiFi and Bluetooth function regardless of
the state of the other phy. This test has some functional differences
from its autotest counterpart.

Instead of expecting a reconnection to an AP when WiFi is enabled, this
test will explicitly ping an AP when WiFi comes up. Additionally,
the bluetooth validation check has been augmented to see if the
adapter can launch a discovery scan properly in order to confirm
functionality of the device.

BUG=b:174159285
TEST=test_that ${DUT} wifi.BluetoothAntennaCoex (board: hatch)

Change-Id: Id414cfb927b865ab8674b6e546937e77a68fc9f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2553205
Reviewed-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>